create interface for router
