+++ internal/models/README.md
@@ -1,4 +1,4 @@
 # Models
 
 # Enchancements
-- Ensure the models does not reflect the data entities fully. The models only contain some fields. This was done due to time limit and ensuring that a minimum viable product can be shown. 
+- Ensure the models does reflect the data entities fully. The models only contain some fields. This was done due to time limit and ensuring that a minimum viable product can be shown.
